Adding support for .yarnrc.yml files.
This change factors commonalities between .npmrc and .yarnrc files into
a common base class `FileProvider`.
This allowed for the addition of .yarnrc.yml files without duplicating code.

This caused some of the logic to move around a bit. The biggest change
for the npmrc is a more genereralized approach to reading, writing and
updating the file without regular expressions and multiple passes.