Add support for passing a config file explicilty.
This is required for larger monorepo's where multiple
languages are mixed, where the .npmrc file is not
necesarily at the root of the workspace.

This PR adds a `-c` or `--config` flag to the cli
to specify the config file to authenticate for.

This will also make it easier to 'bootstrap' and
the readme has been updated with that information