Support native linux usage
azureauth publishes linux deb files since v0.8.2. Add linux/x64 and linux/arm64
to the supported configuration list so that ado-npm-auth can be used there.
Scripts/install.js relies on dpkg-deb to unpack the deb file, which is fine
since the binaries are built for/on Ubuntu where dpkg-deb is guaranteed to be
available. The linux binaries might run on other distros.