ado-npm-auth: Support device code flow
When running headless on linux we can't rely on web flow for authentication.
Add a cli argument that forces azureauth into devicecode flow to support that
scenario. Move from running azureauth with `exec` to `spawn` to make this work.

We need to use spawn so that stderr can be inherited from the calling process.
The device code is displayed to the user on stderr.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>