Expose functionality as a library (#48)
This change exposes the functionality as a library.

It also allows users to pass azureauth location when used as a library when the environment has a custom version of azureauth present already.
Fixes a typo writeWorspaceRegistries => writeWorkspaceRegistries

Improves error handling. In some cases AzureAuth prints warnings on stderr yet still succeeds.